vis weight sum is real, not complex

+++ africanus/averaging/bda_avg.py
@@ -196,7 +196,7 @@ def row_chan_average(meta, flag_row=None, weight=None,
 
         if have_vis:
             vis_avg = np.zeros(out_shape, dtype=vis.dtype)
-            vis_weight_sum = np.zeros(out_shape, dtype=vis.dtype)
+            vis_weight_sum = np.zeros(out_shape, dtype=vis.real.dtype)
             flagged_vis_avg = np.zeros_like(vis_avg)
             flagged_vis_weight_sum = np.zeros_like(vis_weight_sum)
         else:
